requirements.txt
chromadb
python-dotenv
requests
